chore: split auth_5 test into smaller tests
