+++ packages/amplify-e2e-tests/src/__tests__/auth_5e.test.ts
@@ -1,0 +1,64 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import {
+  initJSProjectWithProfile,
+  deleteProject,
+  amplifyPushAuth,
+  updateHeadlessAuth,
+  addAuthWithDefault,
+  createNewProjectDir,
+  deleteProjectDir,
+  getProjectMeta,
+  getUserPool,
+} from '@aws-amplify/amplify-e2e-core';
+import { UpdateAuthRequest } from 'amplify-headless-interface';
+import _ from 'lodash';
+
+const PROJECT_NAME = 'authTest';
+const defaultsSettings = {
+  name: PROJECT_NAME,
+};
+
+describe('headless auth e', () => {
+  let projRoot: string;
+  beforeEach(async () => {
+    projRoot = await createNewProjectDir('auth-update');
+  });
+
+  afterEach(async () => {
+    await deleteProject(projRoot);
+    deleteProjectDir(projRoot);
+  });
+
+  it('updates existing auth resource', async () => {
+    const updateAuthRequest: UpdateAuthRequest = {
+      version: 2,
+      serviceModification: {
+        serviceName: 'Cognito',
+        userPoolModification: {
+          userPoolGroups: [
+            {
+              groupName: 'group1',
+            },
+            {
+              groupName: 'group2',
+            },
+          ],
+        },
+        includeIdentityPool: true,
+        identityPoolModification: {
+          unauthenticatedLogin: true,
+        },
+      },
+    };
+
+    await initJSProjectWithProfile(projRoot, defaultsSettings);
+    await addAuthWithDefault(projRoot, {});
+    await updateHeadlessAuth(projRoot, updateAuthRequest, {});
+    await amplifyPushAuth(projRoot);
+    const meta = getProjectMeta(projRoot);
+    const id = Object.keys(meta.auth).map(key => meta.auth[key])[0].output.UserPoolId;
+    const userPool = await getUserPool(id, meta.providers.awscloudformation.Region);
+    expect(userPool.UserPool).toBeDefined();
+    expect(_.get(meta, ['auth', 'userPoolGroups'])).toBeDefined();
+  });
+});
